#117 - reducing SystemCoordinates intermodule communications class amount
