--- trade-module/src/main/java/io/edpn/backend/trade/application/service/RequestSystemCoordinatesService.java
@@ -1,53 +1,0 @@
-package io.edpn.backend.trade.application.service;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.edpn.backend.messageprocessorlib.application.dto.eddn.data.SystemDataRequest;
-import io.edpn.backend.trade.application.domain.Message;
-import io.edpn.backend.trade.application.domain.System;
-import io.edpn.backend.trade.application.dto.web.object.mapper.MessageMapper;
-import io.edpn.backend.trade.application.port.incomming.kafka.RequestDataUseCase;
-import io.edpn.backend.trade.application.port.outgoing.kafka.SendKafkaMessagePort;
-import io.edpn.backend.trade.application.port.outgoing.systemcoordinaterequest.CreateSystemCoordinateRequestPort;
-import io.edpn.backend.trade.application.port.outgoing.systemcoordinaterequest.ExistsSystemCoordinateRequestPort;
-import io.edpn.backend.util.Module;
-import io.edpn.backend.util.Topic;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-import java.util.Objects;
-
-@RequiredArgsConstructor
-@Slf4j
-public class RequestSystemCoordinatesService implements RequestDataUseCase<System> {
-
-    private final SendKafkaMessagePort sendKafkaMessagePort;
-    private final ExistsSystemCoordinateRequestPort existsSystemCoordinateRequestPort;
-    private final CreateSystemCoordinateRequestPort createSystemCoordinateRequestPort;
-    private final ObjectMapper objectMapper;
-    private final MessageMapper messageMapper;
-
-    @Override
-    public boolean isApplicable(System system) {
-        return Objects.isNull(system.getXCoordinate()) || Objects.isNull(system.getYCoordinate()) || Objects.isNull(system.getZCoordinate());
-    }
-
-    @Override
-    public synchronized void request(System system) {
-        final String systemName = system.getName();
-        boolean shouldRequest = !existsSystemCoordinateRequestPort.exists(systemName);
-        if (shouldRequest) {
-            SystemDataRequest stationDataRequest = new SystemDataRequest(Module.TRADE, systemName);
-
-            JsonNode jsonNode = objectMapper.valueToTree(stationDataRequest);
-
-            Message message = Message.builder()
-                    .topic(Topic.Request.SYSTEM_COORDINATES.getTopicName())
-                    .message(jsonNode.toString())
-                    .build();
-
-            sendKafkaMessagePort.send(messageMapper.map(message));
-            createSystemCoordinateRequestPort.create(systemName);
-        }
-    }
-}
