+++ trade-module/src/test/java/io/edpn/backend/trade/application/service/RequestStationArrivalDistanceServiceTest.java
@@ -71,8 +71,8 @@ public class RequestStationArrivalDistanceServiceTest {
     private RetryTemplate retryTemplate;
     @Mock
     private Executor executor;
-    @Mock
-    private ObjectMapper objectMapper;
+
+    private final ObjectMapper objectMapper = new ObjectMapper();
     @Mock
     private MessageMapper messageMapper;
