+++ trade-module/src/test/java/io/edpn/backend/trade/application/service/RequestSystemCoordinatesServiceTest.java
@@ -6,9 +6,16 @@ import io.edpn.backend.trade.application.domain.System;
 import io.edpn.backend.trade.application.dto.web.object.mapper.MessageMapper;
 import io.edpn.backend.trade.application.port.incomming.kafka.RequestDataUseCase;
 import io.edpn.backend.trade.application.port.outgoing.kafka.SendKafkaMessagePort;
+import io.edpn.backend.trade.application.port.outgoing.system.LoadOrCreateSystemByNamePort;
+import io.edpn.backend.trade.application.port.outgoing.system.LoadSystemsByFilterPort;
+import io.edpn.backend.trade.application.port.outgoing.system.UpdateSystemPort;
 import io.edpn.backend.trade.application.port.outgoing.systemcoordinaterequest.CreateSystemCoordinateRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.systemcoordinaterequest.DeleteSystemCoordinateRequestPort;
 import io.edpn.backend.trade.application.port.outgoing.systemcoordinaterequest.ExistsSystemCoordinateRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.systemcoordinaterequest.LoadAllSystemCoordinateRequestsPort;
 import io.edpn.backend.util.Topic;
+import java.util.concurrent.Executor;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -18,8 +25,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.util.stream.Stream;
+import org.springframework.retry.support.RetryTemplate;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
@@ -30,21 +36,32 @@ import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
 public class RequestSystemCoordinatesServiceTest {
-    
+
     @Mock
-    private SendKafkaMessagePort sendKafkaMessagePort;
-    
+    private LoadSystemsByFilterPort loadSystemsByFilterPort;
+    @Mock
+    private LoadAllSystemCoordinateRequestsPort loadAllSystemCoordinateRequestsPort;
+    @Mock
+    private LoadOrCreateSystemByNamePort loadOrCreateSystemByNamePort;
     @Mock
     private ExistsSystemCoordinateRequestPort existsSystemCoordinateRequestPort;
-    
     @Mock
     private CreateSystemCoordinateRequestPort createSystemCoordinateRequestPort;
-    
-    private final ObjectMapper objectMapper = new ObjectMapper();
-    
+    @Mock
+    private DeleteSystemCoordinateRequestPort deleteSystemCoordinateRequestPort;
+    @Mock
+    private UpdateSystemPort updateSystemPort;
+    @Mock
+    private SendKafkaMessagePort sendKafkaMessagePort;
+    @Mock
+    private RetryTemplate retryTemplate;
+    @Mock
+    private Executor executor;
     @Mock
     private MessageMapper messageMapper;
-    
+
+    private final ObjectMapper objectMapper = new ObjectMapper();
+
     private RequestDataUseCase<System> underTest;
 
     public static Stream<Arguments> provideDoublesForCheckApplicability() {
@@ -60,7 +77,20 @@ public class RequestSystemCoordinatesServiceTest {
 
     @BeforeEach
     void setUp() {
-        underTest = new RequestSystemCoordinatesService(sendKafkaMessagePort, existsSystemCoordinateRequestPort, createSystemCoordinateRequestPort,objectMapper, messageMapper);
+        underTest = new SystemCoordinateInterModuleCommunicationService(
+                loadSystemsByFilterPort,
+                loadAllSystemCoordinateRequestsPort,
+                loadOrCreateSystemByNamePort,
+                existsSystemCoordinateRequestPort,
+                createSystemCoordinateRequestPort,
+                deleteSystemCoordinateRequestPort,
+                updateSystemPort,
+                sendKafkaMessagePort,
+                retryTemplate,
+                executor,
+                objectMapper,
+                messageMapper
+        );
     }
 
     @ParameterizedTest
@@ -82,16 +112,16 @@ public class RequestSystemCoordinatesServiceTest {
                 .build();
 
         underTest.request(system);
-        
+
         ArgumentCaptor<Message> argumentCaptor = ArgumentCaptor.forClass(Message.class);
         verify(messageMapper, times(1)).map(argumentCaptor.capture());
         verify(sendKafkaMessagePort, times(1)).send(messageMapper.map(argumentCaptor.capture()));
-        
+
         Message message = argumentCaptor.getValue();
         assertThat(message, is(notNullValue()));
         assertThat(message.getTopic(), is(Topic.Request.SYSTEM_COORDINATES.getTopicName()));
         assertThat(message.getMessage(), is(notNullValue()));
-        
+
         //TODO: below
         //SystemDataRequest actualSystemDataRequest = objectMapper.treeToValue(message.getMessage(), SystemDataRequest.class);
         assertThat(message.getMessage(), containsString(system.getName()));
